Move all page components into the pages folder
