--- frontend/src/Login.tsx
@@ -1,5 +1,0 @@
-function Login() {
-  return <div>Login</div>;
-}
-
-export default Login;
