+++ frontend/src/pages/App.tsx
@@ -1,15 +1,13 @@
-import Navbar from "./components/Navbar";
-import Calendar from "./components/Calendar";
+import Navbar from "../components/Navbar";
+import Calendar from "../components/Calendar";
 
 function App() {
   return (
     <div className="flex h-screen flex-col overflow-hidden">
-      {/* Give the navbar a fixed track height (tweak as needed) */}
       <header className="h-16 shrink-0">
         <Navbar />
       </header>
 
-      {/* The calendar area gets the remaining height */}
       <main className="flex-1 overflow-hidden">
         <div className="w-[80%] mx-auto h-full p-6">
           <Calendar />
